backend/requirements.txt
fastapi[standard]
sqlmodel
python-dotenv
passlib[bcrypt]
psycopg2-binary
pyjwt
email-validator
password-validator